% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mybrutrarefac.R
\name{My.brut.Rarefac}
\alias{My.brut.Rarefac}
\title{Calcul de la raréfaction}
\usage{
My.brut.Rarefac(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Calcul de la raréfaction
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
