% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mymatrice.R
\name{My.matrice}
\alias{My.matrice}
\title{Créer une matrice a partir d'un jeu données format "long"}
\usage{
My.matrice(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Créer une matrice a partir d'un jeu données format "long"
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
