% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyplotID.R
\name{My.plot.id}
\alias{My.plot.id}
\title{Graphique richesse cumulé & raréfaction x=ID}
\usage{
My.plot.id(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Graphique richesse cumulé & raréfaction x=ID
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
