% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyplotT.R
\name{My.plot.T}
\alias{My.plot.T}
\title{Graphique richesse cumulé & raréfaction x=T}
\usage{
My.plot.T(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Graphique richesse cumulé & raréfaction x=T
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
