% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myall.R
\name{My.all}
\alias{My.all}
\title{Calcul de la richesse cumulé & raréfaction}
\usage{
My.all(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Calcul de la richesse cumulé & raréfaction
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
