% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBrutRichcum.R
\name{My.brut.RichCum}
\alias{My.brut.RichCum}
\title{Calcul de la richesse spécifique cumulé}
\usage{
My.brut.RichCum(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Calcul de la richesse spécifique cumulé
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
