% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myextrapo.R
\name{My.extrapo}
\alias{My.extrapo}
\title{Extrapolation du nombre d'espèces observable}
\usage{
My.extrapo(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{data}

\item{VarSp}{data}

\item{VarT}{data}

\item{Occ}{data}
}
\description{
Extrapolation du nombre d'espèces observable
}
\details{
My.matrice(dataset, VarSp, VarT, Occ)
}
