% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mymatrice.R
\name{My.matrice}
\alias{My.matrice}
\title{Créer une matrice}
\usage{
My.matrice(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{Jeu de données en format long / data.frame}

\item{VarSp}{Variable comprenant les especes / factor}

\item{VarT}{Variable de groupe : session, transects, sites, dates / int}

\item{Occ}{Facultatif : Variable du nombre d'individu observé / int}
}
\value{
data.frame en format "wide"
}
\description{
Créer une matrice -architeture attendu par le package vegan- depuis un jeu de données au format long
}
