% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyplotID.R
\name{My.plot.id}
\alias{My.plot.id}
\title{Graphique de raréfaction et richesse cumulé}
\usage{
My.plot.id(dataset, VarSp, VarT, Occ)
}
\arguments{
\item{dataset}{Jeu de données en format long / data.frame}

\item{VarSp}{Variable comprenant les especes / factor}

\item{VarT}{Variable de groupe : session, transects, sites, dates / int}

\item{Occ}{Facultatif : Variable du nombre d'individu observé / int}
}
\value{
data.frame en format "wide"
}
\description{
Créer un graphique de la raréfaction - calculé à l'aide du package vegan - et de la richesse spécifique cumulé depuis un jeu de données au format long.
x = les identifiants de VarT
}
